Add basic rules cheatsheet
This covers all the core rules other than spacecraft stuff, which
hopefully I will also be able to fit on a single side of A4.  I'm also
planning to make a "references cheatsheet", which'll just have an
index of topics which might come up but are going to be less
common (like poisoning).